+++ lib/src/widgets/layout/platform_tab_scaffold.dart
@@ -39,14 +39,17 @@ class PlatformTabScaffold extends PlatformWidgetKeyedBase {
 
   /// Creates a platform-adaptive tab scaffold.
   ///
-  /// [tabBodyBuilder] is central to determining how state is managed. When it is provided, it is assumed that the state is managed externally
-  /// ([tabDestinations].view is ignored/not needed in this case).
+  /// [tabBodyBuilder] is central to determining how state is managed.
+  /// When it is provided, it is assumed that the state is managed (and rebuilt) externally ([tabDestinations].view is ignored/not needed in this case).
   /// When it is not provided, it is assumed that the state is managed internally and [tabDestinations].view is used to build the tab content.
   ///
   /// Cupertino already has [CupertinoTabView] that makes it trivial to be able to handle state internally.
   /// That is used when [tabDestinations].view is provided.
   ///
-  /// Material revolves around using a `selectedIndex`
+  /// Material revolves around using a `selectedIndex` that is changed when a tab is tapped.
+  ///
+  /// Since [CupertinoTabScaffold].tabBuilder automatically caches the tab content prevent rebuilds when switching tabs,
+  /// the material side has been extended to replicate this functionality.
   const PlatformTabScaffold({
     this.selectedIndex = 0,
     this.tabDestinations,
@@ -240,12 +243,142 @@ class _MaterialTabScaffoldState extends State<_MaterialTabScaffold> {
               },
             ),
           ),
-    body:
-        widget.resolvedTabBodyBuilder?.call(context, widget.resolvedSelectedIndex) ??
-        ValueListenableBuilder(
-          valueListenable: _selectedIndexNotifier!,
-          builder: (_, selectedIndex, _) => widget.resolvedTabDestinations[selectedIndex].view!,
+    body: widget.resolvedTabBodyBuilder != null
+        ? _TabSwitchingView(
+            currentTabIndex: widget.resolvedSelectedIndex,
+            tabCount: widget.resolvedTabDestinations.length,
+            tabBuilder: widget.resolvedTabBodyBuilder!,
+          )
+        : ValueListenableBuilder(
+            valueListenable: _selectedIndexNotifier!,
+            builder: (_, selectedIndex, _) => _TabSwitchingView(
+              currentTabIndex: selectedIndex,
+              tabCount: widget.resolvedTabDestinations.length,
+              tabBuilder: (_, index) => widget.resolvedTabDestinations[index].view!,
+            ),
+          ),
+  );
+}
+
+/// Ripped from Cupertino's `_TabSwitchingView`.
+///
+/// A widget laying out multiple tabs with only one active tab being built
+/// at a time and on stage. Off stage tabs' animations are stopped.
+class _TabSwitchingView extends StatefulWidget {
+  const _TabSwitchingView({
+    required this.currentTabIndex,
+    required this.tabCount,
+    required this.tabBuilder,
+  }) : assert(tabCount > 0, 'Tab count must be greater than 0');
+
+  final int currentTabIndex;
+  final int tabCount;
+  final IndexedWidgetBuilder tabBuilder;
+
+  @override
+  _TabSwitchingViewState createState() => _TabSwitchingViewState();
+}
+
+class _TabSwitchingViewState extends State<_TabSwitchingView> {
+  final shouldBuildTab = <bool>[];
+  final tabFocusNodes = <FocusScopeNode>[];
+
+  // When focus nodes are no longer needed, we need to dispose of them, but we
+  // can't be sure that nothing else is listening to them until this widget is
+  // disposed of, so when they are no longer needed, we move them to this list,
+  // and dispose of them when we dispose of this widget.
+  final discardedNodes = <FocusScopeNode>[];
+
+  @override
+  void initState() {
+    super.initState();
+
+    shouldBuildTab.addAll(List<bool>.filled(widget.tabCount, false));
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+
+    _focusActiveTab();
+  }
+
+  @override
+  void didUpdateWidget(_TabSwitchingView oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    // Only partially invalidate the tabs cache to avoid breaking the current
+    // behavior. We assume that the only possible change is either:
+    // - new tabs are appended to the tab list, or
+    // - some trailing tabs are removed.
+    // If the above assumption is not true, some tabs may lose their state.
+    final lengthDiff = widget.tabCount - shouldBuildTab.length;
+    if (lengthDiff > 0) {
+      shouldBuildTab.addAll(List<bool>.filled(lengthDiff, false));
+    } else if (lengthDiff < 0) {
+      shouldBuildTab.removeRange(widget.tabCount, shouldBuildTab.length);
+    }
+    _focusActiveTab();
+  }
+
+  // Will focus the active tab if the FocusScope above it has focus already. If
+  // not, then it will just mark it as the preferred focus for that scope.
+  void _focusActiveTab() {
+    if (tabFocusNodes.length != widget.tabCount) {
+      if (tabFocusNodes.length > widget.tabCount) {
+        discardedNodes.addAll(tabFocusNodes.sublist(widget.tabCount));
+        tabFocusNodes.removeRange(widget.tabCount, tabFocusNodes.length);
+      } else {
+        tabFocusNodes.addAll(
+          List<FocusScopeNode>.generate(
+            widget.tabCount - tabFocusNodes.length,
+            (index) => FocusScopeNode(
+              debugLabel: '$CupertinoTabScaffold Tab ${index + tabFocusNodes.length}',
+            ),
+          ),
+        );
+      }
+    }
+    FocusScope.of(context).setFirstFocus(tabFocusNodes[widget.currentTabIndex]);
+  }
+
+  @override
+  void dispose() {
+    for (final focusScopeNode in tabFocusNodes) {
+      focusScopeNode.dispose();
+    }
+    for (final focusScopeNode in discardedNodes) {
+      focusScopeNode.dispose();
+    }
+
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) => Stack(
+    fit: StackFit.expand,
+    children: List<Widget>.generate(widget.tabCount, (index) {
+      final active = index == widget.currentTabIndex;
+      shouldBuildTab[index] = active || shouldBuildTab[index];
+
+      return HeroMode(
+        enabled: active,
+        child: Offstage(
+          offstage: !active,
+          child: TickerMode(
+            enabled: active,
+            child: FocusScope(
+              node: tabFocusNodes[index],
+              child: Builder(
+                builder: (context) => shouldBuildTab[index]
+                    ? widget.tabBuilder(context, index)
+                    : const SizedBox.shrink(),
+              ),
+            ),
+          ),
         ),
+      );
+    }),
   );
 }